+++ frontend/javascripts/viewer/model/sagas/saving/rebasing/rebasing_helpers_sagas.ts
@@ -30,7 +30,7 @@ import type {
   UpdateMetadataOfSegmentUpdateAction,
   UpdateSegmentPartialUpdateAction,
   UpdateSegmentVisibilityVolumeAction,
-} from "../volume/update_actions";
+} from "../../volume/update_actions";
 
 export function saveQueueEntriesToServerUpdateActionBatches(
   data: Array<SaveQueueEntry>,
